+++ frontend/src/components/adminDetails/approvalCard/ApprovalCard.tsx
@@ -3,33 +3,31 @@ import CompanyLogo from '../../../assets/Logo Tumbnail.svg';
 import locationSvg from '../assets/Location.svg';
 import style from './approvalCard.module.css';
 import useAdminStore from '../../../store/admin.store';
+import { AdminPropsType } from '../types/types';
+import { useEffect } from 'react';
 
-interface Job {
-  id: number;
-  title: string;
-  description: string;
-  name: string;
-  location: string;
-}
-
-interface ApprovalCardProps {
-  job: Job;
-}
-
-const ApprovalCard: React.FC<ApprovalCardProps> = ({ job }) => {
+const ApprovalCard: React.FC<AdminPropsType> = ({ job }) => {
   const [isClicked, setIsClicked] = useState<boolean>(false);
   const deleteJobRequest = useAdminStore((state) => state.deleteJobRequest);
   const addJobRequest = useAdminStore((state) => state.addJobRequest);
 
   const handleDenyClick = async (id: number) => {
     setIsClicked(true);
-    deleteJobRequest(id);
+    setTimeout(() => {
+      deleteJobRequest(id);
+    }, 300);
   };
   const handleApproveClick = async (id: number) => {
     setIsClicked(true);
-    addJobRequest(id);
+    setTimeout(() => {
+      addJobRequest(id);
+    }, 300);
   };
-
+  useEffect(() => {
+    setTimeout(() => {
+      setIsClicked(false);
+    }, 300);
+  }, [job]);
   return (
     <div className={`${style.cardContainer} ${isClicked ? style.fadeOut : ''}`}>
       <div className={style.logoTitle}>
@@ -38,7 +36,7 @@ const ApprovalCard: React.FC<ApprovalCardProps> = ({ job }) => {
         </div>
         <div className={style.title}>
           <div>{job.title}</div>
-          <div>{job.name}</div>
+          <div className={style.name}>{job.name}</div>
         </div>
       </div>
       <div className={style.buttons}>
@@ -49,7 +47,7 @@ const ApprovalCard: React.FC<ApprovalCardProps> = ({ job }) => {
           className={`${style.btn} ${style.approve}`}
           onClick={() => handleApproveClick(job.id)}
         >
-          Approve
+          Aprove
         </button>
       </div>
       <div className={style.description}>{job.description}</div>
