+++ frontend/src/components/adminDetails/jobApproval/JobApproval.tsx
@@ -1,10 +1,7 @@
 import React from 'react';
 import style from './jobApproval.module.css';
 import ApprovalCard from '../approvalCard/ApprovalCard';
-
-interface JobsProps {
-  jobs: any[];
-}
+import { JobsProps } from '../types/types';
 
 const JobApproval: React.FC<JobsProps> = ({ jobs }) => {
   const jobsToApprove = jobs.slice(0, 4);
@@ -12,8 +9,8 @@ const JobApproval: React.FC<JobsProps> = ({ jobs }) => {
   return (
     <section className={style.content}>
       <div className={style.title}>
-        <h3>need to approve</h3>
-        <p>see all</p>
+        <h3>Need to Approve</h3>
+        <p>See All</p>
       </div>
       <div className={style.cardContent}>
         {jobsToApprove.map((job, index) => (
