+++ frontend/src/pages/admin/AdminPage.tsx
@@ -3,27 +3,10 @@ import useAdminStore from '../../store/admin.store';
 import style from './admin.module.css';
 import CountDisplay from '../../components/adminDetails/countDisplay/CountDisplay';
 import JobApproval from '../../components/adminDetails/jobApproval/JobApproval';
+import { useAdminData } from '../../hooks/UseAdminData';
 
 function AdminPage() {
-  const jobsRequest = useAdminStore((state) => state.jobsRequest);
-  const getJobRequest = useAdminStore((state) => state.getJobRequest);
-  const usersCount = useAdminStore((state) => state.usersCount);
-  const getUserCount = useAdminStore((state) => state.getUserCount);
-  const companiesCount = useAdminStore((state) => state.companiesCount);
-  const getComaniesCount = useAdminStore((state) => state.getComaniesCount);
-  const jobsCount = useAdminStore((state) => state.jobsCount);
-  const getJobsCount = useAdminStore((state) => state.getJobsCount);
-  const openJobsCount = useAdminStore((state) => state.openJobsCount);
-  const getOpenJobsCount = useAdminStore((state) => state.getOpenJobsCount);
-
-  useEffect(() => {
-    getJobRequest();
-    getUserCount();
-    getComaniesCount();
-    getJobsCount();
-    getOpenJobsCount();
-  }, []);
-  console.log(jobsRequest);
+  const { usersCount, companiesCount, jobsCount, jobsRequest, openJobsCount } = useAdminData();
 
   return (
     <div className={style.content}>
