+++ frontend/src/pages/admin/admin.module.css
@@ -2,7 +2,7 @@
   width: 100%;
   height: 100%;
   display: grid;
-  grid-template-rows: 90px 380px auto;
+  grid-template-rows: 90px 400px auto;
   grid-template-columns: auto;
   grid-template-areas: 'top' 'mid' 'btm';
 }
